Fixed wrong input processing variable type

+++ input_processing.py
@@ -108,7 +108,7 @@ class ChessDataProcessor:
         elif field is Fields.LABEL:
             return np.empty(
                 (length,),
-                [("eval", np.int16), ("mate_turns", np.uint8), ("is_mate", np.bool)],
+                [("eval", np.int16), ("mate_turns", np.int16), ("is_mate", np.bool)],
             )
 
         raise ValueError("field should be a singular Fields flag")
